use 64 bit integer (#12)
* Use 64 bit integer for delivery numbers

+++ src/main/java/com/brinkcommerce/api/management/order/delivery/model/response/BrinkDeliveryGetResponse.java
@@ -28,8 +28,8 @@ public record BrinkDeliveryGetResponse(
         Instant updated,
         Instant started,
         Instant restarted,
-        Integer restarts,
+        Long restarts,
         Instant completed,
         Instant failed,
-        Integer revision
+        Long revision
 ) {}
