+++ src/main/java/com/brinkcommerce/api/management/order/delivery/model/response/ShippingFee.java
@@ -4,7 +4,7 @@ package com.brinkcommerce.api.management.order.delivery.model.response;
 public record ShippingFee(
         String shippingFeeId,
         Long taxPercentage,
-        Integer taxPercentageDecimals,
+        Long taxPercentageDecimals,
         String currencyCode,
         Long totalAmount,
         Long totalTaxAmount,
